FiberNlse

FiberNlse.combine — Method

Combines two different Solution structs to an unique one.

FiberNlse.combine — Method

Combines a set of Solution structs to an unique one.

FiberNlse.input — Method

Returns the time and frequency domain input signal of a Solution

FiberNlse.output — Method

Returns the time and frequency domain output signal of a Solution

FiberNlse.raman_linagrawaal — Function

Agrawaal raman model

Default values are (fr = 0.18 , τl = 32e-15 [s], τvib = 12.2e-15 [s])

FiberNlse.RamanModel — Type

RamanModel

Struct describing the raman response of the waveguide

Fields:

fr is the fractional contribution of the Raman effect to the overall nonlienar effect.

time_response is a user defined function returning the Raman impulse response with respect to time.

See also raman_linagrawaal()

FiberNlse.Solution — Type

Solution

Struct containing the result of the simulation.

Fields:

z is the distance vector

t is the duration vector

f is the frequency vector

At is the time-domain propagation field

Af is the frequency-domain (spectrum) propagation field

See also combine()

FiberNlse.Stepper — Type

Stepper

Struct containing the fields that are update or needed during the solver iterations.

FiberNlse.Waveguide — Type

Waveguide

Struct describing the waveguide i.e. the propagation conditions

Fields:

α is the linear optical intensity loss factor (scalar or frequency dependent)

βs is the GVD orders vector (starting at 2nd order)

γ is the nonlinear factor (scalar or frequency dependent)

λc is the center wavelength of the problem (signal and dispersion)

L is the physical length of the waveguide

raman_model is Raman model of the waveguide RamanModel

self_steepening is a boolean to take into account the self self steepening effect

FiberNlse.Waveguide — Method

Constructs a Waveguide struct.

By default no Raman effect nor self steepening is activated.